[FEAT] SharedViewModel 확장 함수 common 모듈로 이동
공통 함수를 사용하는 방식으로 변경
