+++ core/common/src/main/kotlin/com/nexters/ilab/android/core/common/extension/NavBackStackEntry.kt
@@ -1,0 +1,20 @@
+package com.nexters.ilab.android.core.common.extension
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.remember
+import androidx.lifecycle.ViewModel
+import androidx.navigation.NavBackStackEntry
+import androidx.navigation.NavHostController
+import androidx.hilt.navigation.compose.hiltViewModel
+
+// https://youtu.be/h61Wqy3qcKg?si=OqctoATR5MGbypOW
+@Composable
+inline fun <reified T : ViewModel> NavBackStackEntry.sharedViewModel(
+    navController: NavHostController,
+): T {
+    val navGraphRoute = destination.parent?.route ?: return hiltViewModel()
+    val parentEntry = remember(this) {
+        navController.getBackStackEntry(navGraphRoute)
+    }
+    return hiltViewModel(parentEntry)
+}
