+++ feature/createimage/src/main/kotlin/com/nexters/ilab/feature/createimage/navigation/CreateImageNavigation.kt
@@ -1,15 +1,11 @@
 package com.nexters.ilab.feature.createimage.navigation
 
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.ViewModel
-import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
+import com.nexters.ilab.android.core.common.extension.sharedViewModel
 import com.nexters.ilab.feature.createimage.CreateImageCompleteRoute
 import com.nexters.ilab.feature.createimage.CreateImageRoute
 import com.nexters.ilab.feature.createimage.viewmodel.CreateImageViewModel
@@ -52,16 +48,4 @@ fun NavGraphBuilder.createImageNavGraph(
             )
         }
     }
-}
-
-// https://youtu.be/h61Wqy3qcKg?si=OqctoATR5MGbypOW
-@Composable
-inline fun <reified T : ViewModel> NavBackStackEntry.sharedViewModel(
-    navController: NavHostController,
-): T {
-    val navGraphRoute = destination.parent?.route ?: return hiltViewModel()
-    val parentEntry = remember(this) {
-        navController.getBackStackEntry(navGraphRoute)
-    }
-    return hiltViewModel(parentEntry)
 }
