+++ feature/mypage/src/main/kotlin/com/nexters/ilab/android/feature/mypage/navigation/MyPageNavigation.kt
@@ -1,17 +1,13 @@
 package com.nexters.ilab.android.feature.mypage.navigation
 
 import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.ViewModel
-import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.navigation
+import com.nexters.ilab.android.core.common.extension.sharedViewModel
 import com.nexters.ilab.android.feature.mypage.MyAlbumImageRoute
 import com.nexters.ilab.android.feature.mypage.MyPageRoute
 import com.nexters.ilab.android.feature.mypage.MyPageViewModel
@@ -59,15 +55,4 @@ fun NavGraphBuilder.myPageNavGraph(
             )
         }
     }
-}
-
-@Composable
-inline fun <reified T : ViewModel> NavBackStackEntry.sharedViewModel(
-    navController: NavHostController,
-): T {
-    val navGraphRoute = destination.parent?.route ?: return hiltViewModel()
-    val parentEntry = remember(this) {
-        navController.getBackStackEntry(navGraphRoute)
-    }
-    return hiltViewModel(parentEntry)
 }
