+++ feature/uploadphoto/src/main/kotlin/com/nexters/ilab/android/feature/uploadphoto/navigation/UploadPhotoNavigation.kt
@@ -1,16 +1,12 @@
 package com.nexters.ilab.android.feature.uploadphoto.navigation
 
-import androidx.compose.runtime.Composable
-import androidx.compose.runtime.remember
-import androidx.hilt.navigation.compose.hiltViewModel
-import androidx.lifecycle.ViewModel
-import androidx.navigation.NavBackStackEntry
 import androidx.navigation.NavController
 import androidx.navigation.NavGraphBuilder
 import androidx.navigation.NavHostController
 import androidx.navigation.NavOptions
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.navigation
+import com.nexters.ilab.android.core.common.extension.sharedViewModel
 import com.nexters.ilab.android.feature.uploadphoto.InputStyleRoute
 import com.nexters.ilab.android.feature.uploadphoto.UploadCheckRoute
 import com.nexters.ilab.android.feature.uploadphoto.UploadPhotoRoute
@@ -74,16 +70,4 @@ fun NavGraphBuilder.uploadPhotoNavGraph(
             )
         }
     }
-}
-
-// https://youtu.be/h61Wqy3qcKg?si=OqctoATR5MGbypOW
-@Composable
-inline fun <reified T : ViewModel> NavBackStackEntry.sharedViewModel(
-    navController: NavHostController,
-): T {
-    val navGraphRoute = destination.parent?.route ?: return hiltViewModel()
-    val parentEntry = remember(this) {
-        navController.getBackStackEntry(navGraphRoute)
-    }
-    return hiltViewModel(parentEntry)
 }
